changes:
-  ui redesigned.

+++ src/components/SingleBlog.vue
@@ -1,5 +1,5 @@
 <template>
-  <div class="container">
+  <b-container>
     <b-form class="my-style" v-if="!show">
       <b-form-group id="input-group-1" label="Blog Title:" label-for="input-1">
         <b-form-input id="input-1" v-model="blog.title" type="text" placeholder="Blog Title"></b-form-input>
@@ -14,7 +14,7 @@
           class="mt-3 mb-3"
           accept="image/*"
           plain
-          placeholder="select add Image"
+          placeholder="add Image"
           v-if="!flag"
         ></b-form-file>
         <b-progress class="mb-2" variant="primary" :value="value" :max="max" v-if="!flag"></b-progress>
@@ -34,7 +34,6 @@
       </b-form-group>
       <b-form-select class="mb-2" v-model="blog.author" :options="authors"></b-form-select>
       <b-button class="mr-2 mt-2" variant="primary" v-on:click="update">Update</b-button>
-      <b-button class="mr-2 mt-2" variant="primary" v-on:click="removeBlog">Remove</b-button>
       <b-button class="mt-2" variant="primary" v-on:click="cancel">Cancel</b-button>
     </b-form>
     <div class="my-style" v-if="show">
@@ -53,10 +52,11 @@
       <br />
       <p>Author: {{blog.author}}</p>
       <br />
-      <b-button class="mt-2 mr-2" variant="primary" v-on:click="toggle">To Update</b-button>
+      <b-button class="mt-2 mr-2" variant="primary" v-on:click="toggle">Edit Blog</b-button>
+      <b-button class="mr-2 mt-2" variant="primary" v-on:click="removeBlog">Remove Blog</b-button>
       <b-button class="mt-2" variant="primary" v-on:click="cancel">Cancel</b-button>
     </div>
-  </div>
+  </b-container>
 </template>
 
 <script>
@@ -72,7 +72,7 @@ export default {
       file: null,
       value: 0,
       max: 100,
-      flag: true
+      flag: true,
     };
   },
   watch: {
@@ -82,64 +82,64 @@ export default {
       var uploadTask = sr.put(this.file);
       uploadTask.on(
         "state_changed",
-        snapshot => {
+        (snapshot) => {
           this.value =
             (snapshot.bytesTransferred / snapshot.totalBytes) * this.max;
         },
         null,
         () => {
-          uploadTask.snapshot.ref.getDownloadURL().then(url => {
+          uploadTask.snapshot.ref.getDownloadURL().then((url) => {
             this.blog.remoteImgUrl = url;
           });
         }
       );
-    }
+    },
   },
   computed: {
     blog() {
       return this.$store.getters.getBlogById(this.id);
-    }
+    },
   },
   methods: {
     ...mapActions(["deleteBlog", "updateBlog"]),
-    update: function() {
+    update: function () {
       this.updateBlog({
         id: this.id,
         title: this.blog.title,
         content: this.blog.content,
         remoteImgUrl: this.blog.remoteImgUrl,
         categories: this.blog.categories,
-        author: this.blog.author
+        author: this.blog.author,
       });
       this.$router.push("/");
     },
-    removeBlog: function() {
-      if (this.blog.remoteImgUrl === "") {
+    removeBlog: function () {
+      const answerFlag = confirm("sind Sie sicher?");
+      if (answerFlag) {
+        if (this.blog.remoteImgUrl === "") {
+          this.deleteBlog(this.id);
+          this.$router.push("/");
+        }
         this.deleteBlog(this.id);
+        this.removeImage();
         this.$router.push("/");
       }
-      this.deleteBlog(this.id);
-      this.removeImage();
-      this.$router.push("/");
     },
-    removeImage: function() {
+    removeImage: function () {
       var desertRef = storageRef.refFromURL(this.blog.remoteImgUrl);
       desertRef.delete().then(() => {
         this.flag = !this.flag;
       });
     },
-    toggle: function() {
+    toggle: function () {
       this.show = !this.show;
     },
-    cancel: function() {
+    cancel: function () {
       this.$router.push("/");
-    }
-  }
+    },
+  },
 };
 </script>
 
 <style scoped lang="scss">
-.my-mb {
-  margin-bottom: 15px;
-}
 </style>
